requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
pydantic>=1.10,<3.0
PyYAML>=6.0,<7.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib>=1.7,<2.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
numpy>=1.24,<3.0
